Change propertyName to propertyDescriptor
This is a breaking change.
Change the storage interface to use a PropertyDescriptor type which
includes both a property name and a namespace.  This prevents property
name collisions which will occur in the SPDX v3 model update.

To update any code which uses the storage interface, change the property
name to PropertyDescriptor for all calls to the interface which used
propertyName.

The constants have been updated to use the PropertyDescriptor type which
should minimize the amount of change.

Signed-off-by: Gary O'Neall <gary@sourceauditor.com>