Remove the patch version on the license list
The patch version was added in the most recent release of the SPDX
license list to be compatible with the SPDX 3.0 release.  This, however,
broke the validation for the SPDX 2.X parser which checks that there
is no patch version (per the 2.X spec).  This commit removes the patch
version from the license list version of the license list for the SPDX
since this version of the license list only supports version 2.X

Signed-off-by: Gary O'Neall <gary@sourceauditor.com>