Change logging for license list I/O error
Changes the logging level from error to warn if the license list can not
be accessed over the network.  The logic being that the impact is not
substantial since you may only be referencing a stale version of the
license list rather than the most recent.

The warning message was also enhanced to point to the
useJARLicenseInfoOnly property.

Reference related issue
https://github.com/spdx/spdx-maven-plugin/issues/172