Add creationInfo during license parsing
Previous SPDX version 3 license parsing results produced invalid
license due to missing creation informations.

This PR introduces a potential incompatibility by requiring a
creationInfo parameter to the license parser.  This was intentional so
that any callers of the method would be required to consider the correct
creationInfo to be used.

Signed-off-by: Gary O'Neall <gary@sourceauditor.com>